12. Implemented ChartService Interface. Getting ready for mock data.
