+++ Spring/Project/anestheticChart/src/main/java/lt/codeacademy/anestheticChart/service/ChartService.java
@@ -1,23 +1,10 @@
 package lt.codeacademy.anestheticChart.service;
 
-import lombok.RequiredArgsConstructor;
 import lt.codeacademy.anestheticChart.model.ChartDTO;
-import lt.codeacademy.anestheticChart.repository.ChartRepository;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-@RequiredArgsConstructor
-@Service
-public class ChartService {
-
-    private final ChartRepository chartRepository;
-
-    public void addCharts(ChartDTO chartDTO){
-        chartRepository.save(chartDTO);
-    }
-
-    public List<ChartDTO> getCharts() {
-        return chartRepository.getCharts();
-    }
+public interface ChartService {
+    void addCharts(ChartDTO chartDTO);
+    List<ChartDTO> getCharts();
 }
